Remove unneeded toString(QPair) declaration
It's literally the same as the std::pair one that immediately follows
it, because QPair these days is just a template alias for std::pair.

Wasn't flagged as an error because it's just a declaration, not a
definition.

Pick-to: 6.5
Change-Id: I5b47572f6e2a71edb47cbe224801a719ff1e060d
Reviewed-by: Jason McDonald <macadder1@gmail.com>
(cherry picked from commit 85842da95b5a7821b44d83e5d26922a3d07ef6cf)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>