qcompare.h: move QPartialOrdering to the end of the header
This makes sure that everything in-between will not accidentally come
to depend on QPartialOrdering. It also makes it much easier to
centralize the conversions to/from Qt::_ordering types in the
QPartialOrdering class, keeping the newfangled clases clean of legacy.

I should have done it this way to begin with... Mea culpa.

Amends 4b6f757020382ed157bf6beb572549f05e881359.

Change-Id: I269bdf0cbde59d317381ccd06001e56d5b3c289a
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>