Doc: Add AcceptMode info to DontConfirmOverwrite
When QFileDialog::DontConfirmOverwrite is in the false state (default state),
it requires QFileDialog::AcceptMode to be set to QFileDialog::AcceptSave.
This information was not stated in the documents and this fix adds this
information to the docs.

Fixes: QTBUG-49720
Pick-to: 6.6 6.5
Change-Id: I1017252980810e712f2710751c8852dd12eaeef7
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
(cherry picked from commit 76a1c7690fd9f8e747a2bbdd5d910bebd6ef162c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>