QByteArray: inline QByteArray::indexOf() and use a char overload
QByteArray::indexOf() had the best implementation of the two *because*
it was out-of-line, in qbytearray.cpp. The compiler could thus see the
body of the QtPrivate::findByteArray() function and inline it, with
constant-propagation, so only findCharHelper() was expanded and we had a
fast path towards memchr().

On the other hand, QByteArrayView::indexOf() was inline, so the
compiler emitted the call to QtPrivate::findByteArray() in user code,
causing the full function to be executed. We fix that by adding a char
overload for QtPrivate::findByteArray() and lastIndexOf(), so we get to
memchr() more quickly.

Also, inline QByteArray::indexOf() and lastIndexOf().

Before:
QByteArray::indexOf():
     8.83287376 nsecs per iteration
     22.01766832 CPU cycles per iteration, 2.49 GHz
     62.00000330 instructions per iteration, 2.816 instr/cycle
     21.00000281 branch instructions per iteration, 2.38 G/sec
QByteArrayView::indexOf():
     9.64034694 nsecs per iteration
     24.03001151 CPU cycles per iteration, 2.49 GHz
     68.00000355 instructions per iteration, 2.830 instr/cycle
     23.00000306 branch instructions per iteration, 2.39 G/sec

After (same result for both, requires recompilation):
     8.83207052 nsecs per iteration
     22.01568546 CPU cycles per iteration, 2.49 GHz
     60.00000331 instructions per iteration, 2.725 instr/cycle
     21.00000281 branch instructions per iteration, 2.38 G/sec

The inlining appears to have cut 2 instructions, but the effect is not
measurable in this benchmark (the entire code is in cache and the Branch
Predictor is probably primed).

Fixes: QTBUG-119750
Change-Id: Ica7a43f6147b49c187ccfffd179e1cb4b306fc62
Reviewed-by: Marc Mutz <marc.mutz@qt.io>