wasm: delay "qtLoaded" qtloader.js event
Currently we send qtLoaded in response to Emscripten
onRuntimeInitialized, which is sent just before or at
the time main() is called.

This can be too early if Qt (app) initialization is
not instantaneous, in which case we risk displaying
a blank page in between hiding the loading screen
and showing the application.

Change this to send qtLoaded when QCoreApplication has
finished its constructor and the event loop is running.
Also add the possibility of registering startup tasks
which can delay sending qtLoaded further.

We now require a QCoreApplication (or a subclass) instance
for the qtLoaded event. onRuntimeInitialized can be used
as a replacement if anyone needs the previous behavior.

Change-Id: I9298e881d2909aac4040c5f2068e6c7d196196cc
Reviewed-by: Lorn Potter <lorn.potter@gmail.com>